# docs/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "BankCards API",
    "description": "REST API для управления банковскими картами",
    "contact": {
      "name": "Kitoglav",
      "email": "vlad_lanayev@vk.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "Bearer Authentication": []
    }
  ],
...
  "paths": {
...
    "/api/users": {
      "get": {
        "tags": [
          "Пользователи"
        ],
        "summary": "Получить информацию о текущем пользователе",
        "description": "Возвращает информацию о текущем аутентифицированном пользователе",
        "operationId": "getUser",
        "responses": {
          "200": {
            "description": "Информация получена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/users/{userId}": {
      "get": {
        "tags": [
          "Администрирование пользователей"
        ],
        "summary": "Получить пользователя по ID",
        "description": "Возвращает информацию о пользователе по его идентификатору",
        "operationId": "getUser_1",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID пользователя",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Пользователь найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "404": {
            "description": "Пользователь не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Администрирование пользователей"
        ],
        "summary": "Удалить пользователя",
        "description": "Удаляет пользователя из системы",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID пользователя",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Пользователь удален",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "204": {
            "description": "Пользователь не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "409": {
            "description": "У пользователя есть активные карты",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer Authentication": {
        "type": "http",
        "description": "Введите JWT токен без префикса 'Bearer'",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}